third_party/blink/web_tests/shapedetection/detection-HTMLImageElement.html

<!DOCTYPE html>
<script src="../resources/testharness.js"></script>
<script src="../resources/testharnessreport.js"></script>
<script src="file:///gen/layout_test_data/mojo/public/js/mojo_bindings.js"></script>
<script src="file:///gen/mojo/public/mojom/base/big_buffer.mojom.js"></script>
<script src="file:///gen/services/shape_detection/public/mojom/textdetection.mojom.js"></script>
<script src="resources/big-buffer-helpers.js"></script>
<script src="resources/mock-textdetection.js"></script>
<body>
<img id="img" src="../media/content/greenbox.png"/>
</body>
<script>

var createTestForImageElement = function(createDetector, mock,
                                         detectionResultTest) {
  promise_test(async function() {
    var img = document.getElementById("img");

    var detector = createDetector();
    try {
      var detectionResult = await detector.detect(img);
      detectionResultTest(detectionResult, mock);
    } catch(error) {
      assert_unreached("Error during detect(img): " + error);
    }
  });
};

function TextDetectorDetectionResultTest(detectionResult, mock) {
  assert_equals(detectionResult.length, 2, "Number of textBlocks");
  assert_equals(detectionResult[0].rawValue, "cats", "textBlock 1");
  assert_equals(detectionResult[1].rawValue, "dogs", "textBlock 2");
  for (i = 0; i < detectionResult.length; i++) {
    assert_equals(detectionResult[i].boundingBox.x, detectionResult[i].cornerPoints[0].x);
    assert_equals(detectionResult[i].boundingBox.y, detectionResult[i].cornerPoints[0].y);
    assert_equals(detectionResult[i].boundingBox.width,
                  detectionResult[i].cornerPoints[2].x - detectionResult[i].cornerPoints[3].x);
    assert_equals(detectionResult[i].boundingBox.height,
                  detectionResult[i].cornerPoints[2].y - detectionResult[i].cornerPoints[1].y);
  }
}

// These tests verify that a Detector's detect() works on an HTMLImageElement.
// Use the mock mojo server implemented in mock-textdetection.js.
generate_tests(createTestForImageElement, [
  [
    "Text - detect(HTMLImageElement)",
    () => { return new TextDetector(); },
    mockTextDetection,
    TextDetectorDetectionResultTest
  ]
]);

</script>
